refactor(api-gateway): response object modified and used
