+++ API-Gateway/src/routes/response/index.js
@@ -1,0 +1,16 @@
+export default class Response {
+  set statusCode(code) {
+    this._statusCode = code;
+  }
+
+  set data(responseData) {
+    this._data = responseData;
+  }
+
+  toJSON() {
+    return { 
+      statusCode: this._statusCode,
+      data: this._data,
+    };
+  }
+}
