--- API-Gateway/src/routes/response/response.js
@@ -1,7 +1,0 @@
-const response = function() {
-  return {
-    statusCode: '',
-    data: {},
-  };
-};
-module.exports = response;
